requirements.txt
pygame
pytorch
matplotlib